requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-core
langsmith
unstructured
chromadb
openai
tiktoken
flask
gunicorn
streamlit